fix(wextra): fix -Wextra warnings in data_format, dispatch, kernels, profiler
- bfloat{2,4,8,16}.cpp, blockfloat_common.cpp: cast uint32_t limits in int loops
- distributed_tensor_configs.cpp: size_t i for .size() loops
- simple_trace_allocator.cpp: add missing .size=0,.sync_count=0 to ConfigBufferEntry inits
- ringbuffer_cache.cpp: uint32_t required_space (was int vs uint32_t cache_size)
- fd_kernel.cpp: fix corrupted DataMovementConfig (duplicate defines field), add named_compile_args to EthernetConfig
- kernel.cpp: add .math_fidelity=nullopt to KernelMeta init, cast index in TT_ASSERTs
- free_list_opt.cpp: cast size_t limits, fix sentinel comparisons
- dispatch.cpp, dispatch_s.cpp: static_cast<int>() for get_programmable_core_type_index
- relay_mux.cpp: uint32_t i for fabric_channels loop
- program.cpp: fix misplaced static_cast<int>() wrapping
- noc_event_profiler_utils.hpp: uint32_t eth_chan
- tt_metal_profiler.cpp: add named_compile_args, fix decltype cast to uint64_t

+++ tt_metal/impl/data_format/bfloat8.cpp
@@ -116,12 +116,12 @@ std::vector<float> unpack_bfp8_tiles_into_float_vec(
     uint32_t fp32_element_index = 0;
     std::vector<float> float_vec;
     float_vec.resize(num_tiles * num_float_in_tile);
-    for (int tile_index = 0; tile_index < num_tiles; ++tile_index) {
-        for (int tr = 0; tr < subtiles_in_tile_row; ++tr) {
-            for (int tc = 0; tc < subtiles_in_tile_col; ++tc) {
-                for (int i = 0; i < subtile_rows; ++i) {
+    for (int tile_index = 0; tile_index < static_cast<int>(num_tiles); ++tile_index) {
+        for (int tr = 0; tr < static_cast<int>(subtiles_in_tile_row); ++tr) {
+            for (int tc = 0; tc < static_cast<int>(subtiles_in_tile_col); ++tc) {
+                for (int i = 0; i < static_cast<int>(subtile_rows); ++i) {
                     subtile_r = tr * face_H + i;
-                    for (int j = 0; j < subtile_cols; j += 8) {
+                    for (int j = 0; j < static_cast<int>(subtile_cols); j += 8) {
                         subtile_c = tc * face_W + j;
                         data_index =
                             (tr * (subtiles_in_tile_col * face_HW / 4) + tc * (face_HW / 4) + i * (face_W / 4) +
