+++ tt_metal/impl/data_format/blockfloat_common.cpp
@@ -382,16 +382,16 @@ std::vector<uint32_t> pack_as_bfp_tiles(
         num_mantissas_in_dword = 4;
     }
     int fp32_element_index = 0;
-    for (int tile_index = 0; tile_index < num_tiles; ++tile_index) {
+    for (int tile_index = 0; tile_index < static_cast<int>(num_tiles); ++tile_index) {
         std::vector<uint32_t> packed_data;
         std::vector<uint8_t> exponents_with_padding;
         exponents_with_padding.reserve(l1_alignment * subtiles_in_tile_row * subtiles_in_tile_col);
-        for (int tr = 0; tr < subtiles_in_tile_row; ++tr) {
-            for (int tc = 0; tc < subtiles_in_tile_col; ++tc) {
-                for (int i = 0; i < subtile_rows; ++i) {
+        for (int tr = 0; tr < static_cast<int>(subtiles_in_tile_row); ++tr) {
+            for (int tc = 0; tc < static_cast<int>(subtiles_in_tile_col); ++tc) {
+                for (int i = 0; i < static_cast<int>(subtile_rows); ++i) {
                     std::vector<uint32_t> single_row;
                     // populate a single row
-                    for (int j = 0; j < subtile_cols; ++j) {
+                    for (int j = 0; j < static_cast<int>(subtile_cols); ++j) {
                         int data_index;
                         if (row_major_input) {
                             data_index =
